Renamed gridfunctions file to hephaestus_containers to be consistent and moved into named_fields_map directory.
