+++ src/auxsolvers/auxsolver_base.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include "coefficients.hpp"
-#include "gridfunctions.hpp"
+#include "hephaestus_containers.hpp"
 #include "hephaestus_solvers.hpp"
 #include "inputs.hpp"
 #include "mfem.hpp"
