--- src/gridfunctions/CMakeLists.txt
@@ -1,1 +1,0 @@
-file(GLOB_RECURSE src_files "*.h" "*.hpp" "*.cpp")
